Update egg1-handson using Cloud Run
