+++ gaming/egg1/README.md
@@ -1,4 +1,4 @@
-# GCP Handson materials for EGG #1 
+# GCP Handson materials for EGG #1
 
 [![Open in Cloud Shell](https://gstatic.com/cloudssh/images/open-btn.png)](https://ssh.cloud.google.com/cloudshell/open?cloudshell_git_repo=https://github.com/GoogleCloudPlatform/gcp-getting-started-lab-jp&cloudshell_working_dir=gaming/egg1&cloudshell_tutorial=tutorial.md)
