--- gaming/egg1/answer/app.yaml
@@ -1,17 +1,0 @@
-runtime: go112
-
-handlers:
-  - url: /.*
-    secure: always
-    script: auto
-
-vpc_access_connector:
-  name: "projects/your-project-id/locations/us-central1/connectors/egg-vpc-connector"
-
-env_variables:
-  DB_INSTANCE: "your-project-id:us-central1:eggsql-1"
-  DB_USER: root
-  DB_PASS: xxxxxx
-  REDIS_HOST: 0.0.0.0
-  REDIS_PORT: 6379
-
