--- gaming/egg1/answer/go.mod
@@ -1,15 +1,0 @@
-module egg1
-
-go 1.12
-
-require (
-	cloud.google.com/go/firestore v1.2.0
-	github.com/c2h5oh/datasize v0.0.0-20200112174442-28bbd4740fee // indirect
-	github.com/go-sql-driver/mysql v1.5.0
-	github.com/gomodule/redigo v2.0.0+incompatible
-	github.com/influxdata/tdigest v0.0.1 // indirect
-	github.com/mailru/easyjson v0.7.1 // indirect
-	github.com/tsenart/go-tsz v0.0.0-20180814235614-0bd30b3df1c3 // indirect
-	github.com/tsenart/vegeta v12.7.0+incompatible // indirect
-	google.golang.org/api v0.21.0
-)
