+++ gaming/egg1/go.mod
@@ -1,0 +1,15 @@
+module github.com/GoogleCloudPlatform/gcp-getting-started-lab-jp/gaming/egg1/
+
+go 1.15
+
+require (
+	cloud.google.com/go/firestore v1.3.0
+	github.com/gomodule/redigo v2.0.0+incompatible
+	golang.org/x/net v0.0.0-20200904194848-62affa334b73 // indirect
+	golang.org/x/oauth2 v0.0.0-20200902213428-5d25da1a8d43 // indirect
+	golang.org/x/sys v0.0.0-20200909081042-eff7692f9009 // indirect
+	golang.org/x/tools v0.0.0-20200908211811-12e1bf57a112 // indirect
+	google.golang.org/api v0.31.0
+	google.golang.org/genproto v0.0.0-20200904004341-0bd0a958aa1d // indirect
+	google.golang.org/grpc v1.32.0 // indirect
+)
